+++ empirical/scripts/emp_aggregation.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python2
+#!/usr/bin/env python3
 """Script to aggregate individual IM csv files into a single file.
 
 Sites in the individual IMs files have to match.
@@ -6,6 +6,7 @@ Resulting csv file will have consistent column order.
 """
 import argparse
 import os
+
 import numpy as np
 import pandas as pd
 
@@ -17,31 +18,8 @@ STATION_COL_NAME = "station"
 COMPONENT_COL_NAME = "component"
 
 
-def aggregate_data():
-    parser = argparse.ArgumentParser(
-        description="Script to aggregate individual IM " "csv files into a single file."
-    )
-
-    parser.add_argument(
-        "-o",
-        "--output_dir",
-        required=True,
-        help="path to output folder that stores the " "aggregated measures",
-    )
-    parser.add_argument("-i", "--identifier", help="run-name for run")
-    parser.add_argument(
-        "-r",
-        "--rupture",
-        default="unknown",
-        help="Please specify the rupture name of the " "simulation. eg.Albury",
-    )
-    parser.add_argument(
-        "-v", "--version", default="XXpY", help="The version of the simulation. eg.18p4"
-    )
-    parser.add_argument("im_files", nargs="+")
-    args = parser.parse_args()
-
-    dfs = [pd.read_csv(im_file) for im_file in args.im_files]
+def aggregate_data(im_files, output_dir, identifier, rupture, version):
+    dfs = [pd.read_csv(im_file) for im_file in im_files]
 
     # Check that they all have the same number of rows
     # and that the stations match
@@ -57,7 +35,7 @@ def aggregate_data():
                 raise Exception(
                     "Input files {} and {} are incompatible. Either different "
                     "number of entries, or stations/sites don't match".format(
-                        args.im_files[0], args.im_files[ix]
+                        im_files[0], im_files[ix]
                     )
                 )
 
@@ -78,24 +56,52 @@ def aggregate_data():
     result_df = order_im_cols_df(result_df)
 
     result_df.to_csv(
-        os.path.join(args.output_dir, "{}.csv".format(args.identifier)),
+        os.path.join(output_dir, "{}.csv".format(identifier)),
         index=False,
         float_format="%.6f",
     )
 
     # Metadata file
-    metadata_fname = "{}_empirical.info".format(args.identifier)
-    metadata_path = os.path.join(args.output_dir, metadata_fname)
+    metadata_fname = "{}_empirical.info".format(identifier)
+    metadata_path = os.path.join(output_dir, metadata_fname)
     with open(metadata_path, "w") as f:
         date = datetime.now().strftime("%Y%m%d_%H%M%S")
 
         f.write("identifier,rupture,type,date,version\n")
-        f.write(
-            "{},{},empirical,{},{}".format(
-                args.identifier, args.rupture, date, args.version
-            )
-        )
+        f.write("{},{},empirical,{},{}".format(identifier, rupture, date, version))
+
+
+def load_args():
+    parser = argparse.ArgumentParser(
+        description="Script to aggregate individual IM " "csv files into a single file."
+    )
+    parser.add_argument(
+        "-o",
+        "--output_dir",
+        required=True,
+        help="path to output folder that stores the " "aggregated measures",
+    )
+    parser.add_argument("-i", "--identifier", help="run-name for run")
+    parser.add_argument(
+        "-r",
+        "--rupture",
+        default="unknown",
+        help="Please specify the rupture name of the " "simulation. eg.Albury",
+    )
+    parser.add_argument(
+        "-v", "--version", default="XXpY", help="The version of the simulation. eg.18p4"
+    )
+    parser.add_argument("im_files", nargs="+")
+    return parser.parse_args()
+
+
+def main():
+    args = load_args()
+
+    aggregate_data(
+        args.im_files, args.output_dir, args.identifier, args.rupture, args.version
+    )
 
 
 if __name__ == "__main__":
-    aggregate_data()
+    main()